requirements.txt
openai>=1.0.0
pillow>=10.0.0
numpy>=1.25.0